pyproject.toml
[tool.poetry]
name = "bookslibrary"
version = "0.1.0"
description = ""
authors = ["Evgeny Eroshin <eee333@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
